v0.7.0: backend refactoring
